refactor(projectListMfe): reorganize portfolio components into page-b… …ased architecture
- Move portfolio components from flat structure to pages/Portfolio directory
- Extract portfolio logic from app.tsx into dedicated Portfolio page component
- Update TypeScript configurations to support new structure
- Maintain all existing functionality while improving code organization
- Add routing support for multiple portfolio views (Classic and Creative)
