+++ apps/projectListMfe/src/app/pages/Portfolio/Portfolio.scss
@@ -1,0 +1,111 @@
+.portfolio-nav {
+  position: sticky;
+  top: 0;
+  z-index: 1000;
+  background: rgba(10, 10, 10, 0.95);
+  backdrop-filter: blur(20px);
+  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+
+  .nav-container {
+    margin: 0 auto;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 1rem 2rem;
+  }
+
+  .nav-logo {
+    text-decoration: none;
+    
+    .logo-text {
+      font-size: 1.5rem;
+      font-weight: 700;
+      // background: linear-gradient(135deg, #00ffff, #ff00ff);
+      -webkit-background-clip: text;
+      background-clip: text;
+      -webkit-text-fill-color: transparent;
+    }
+  }
+
+  .nav-links {
+    display: flex;
+    gap: 2rem;
+
+    .nav-link {
+      text-decoration: none;
+      color: rgba(255, 255, 255, 0.7);
+      font-weight: 500;
+      padding: 0.5rem 1rem;
+      border-radius: 8px;
+      transition: all 0.3s ease;
+      position: relative;
+
+      &:hover {
+        color: #00ffff;
+        background: rgba(255, 255, 255, 0.05);
+      }
+
+      &.active {
+        color: #00ffff;
+        background: rgba(0, 255, 255, 0.1);
+        
+        &::after {
+          content: '';
+          position: absolute;
+          bottom: -1rem;
+          left: 50%;
+          transform: translateX(-50%);
+          width: 6px;
+          height: 6px;
+          background: #00ffff;
+          border-radius: 50%;
+          box-shadow: 0 0 10px #00ffff;
+        }
+      }
+    }
+  }
+}
+
+.wrapper {
+  padding: 2rem;
+  margin: 0 auto;
+  min-height: calc(100vh - 70px);
+}
+
+.main-title {
+  border-bottom: 3px solid #eee;
+  color: #2a3d45;
+  font-size: 2.5rem;
+  margin-bottom: 2rem;
+  padding-bottom: 1rem;
+}
+
+.no-results {
+  text-align: center;
+  padding: 4rem 2rem;
+  color: rgba(255, 255, 255, 0.7);
+
+  h3 {
+    font-size: 1.5rem;
+    margin-bottom: 0.5rem;
+    color: white;
+  }
+
+  p {
+    font-size: 1rem;
+  }
+}
+
+@media (max-width: 768px) {
+  .portfolio-nav .nav-container {
+    padding: 1rem;
+  }
+  
+  .portfolio-nav .nav-links {
+    gap: 1rem;
+  }
+  
+  .wrapper {
+    padding: 1rem;
+  }
+}
