+++ apps/projectListMfe/src/app/pages/Portfolio/components/Filter/HighlightedFilters.scss
