+++ apps/projectListMfe/src/app/pages/Portfolio/components/Project.scss
