+++ apps/projectListMfe/tsconfig.app.json
@@ -31,10 +31,10 @@
   "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx"],
   "references": [
     {
-      "path": "../../libs/portfolio-store/tsconfig.lib.json"
+      "path": "../../libs/store/tsconfig.lib.json"
     },
     {
-      "path": "../../libs/store/tsconfig.lib.json"
+      "path": "../../libs/portfolio-store/tsconfig.lib.json"
     }
   ]
 }
