+++ apps/projectListMfe/tsconfig.json
@@ -3,10 +3,10 @@
   "include": [],
   "references": [
     {
-      "path": "../../libs/portfolio-store"
+      "path": "../../libs/store"
     },
     {
-      "path": "../../libs/store"
+      "path": "../../libs/portfolio-store"
     },
     {
       "path": "./tsconfig.app.json"
