+++ apps/projectListMfe/tsconfig.spec.json
@@ -28,6 +28,9 @@
       "path": "./tsconfig.app.json"
     },
     {
+      "path": "../../libs/shared/config/tsconfig.lib.json"
+    },
+    {
       "path": "../../libs/store/tsconfig.lib.json"
     },
     {
